fgdb version 0.5
- Removed edit session, it was doubling the processing time. The insert locks 
seem to be related to using arcpy to check if a fgdb exists and delete it. 
Changed method to os.path.isdir() and shutil.rmtree() and have not been able 
to replicate cursor lock errors. If these lock errors resurface, will explore
arcpy.management.ClearWorkspaceCache() or providing option create with Edit
session. 
- improved messaging
- Added some gc.collect() arguments